# backend/swagger.yaml
...
openapi: 3.1.0
info:
  title: VisionSea MCP Backend
  version: 1.0.0
paths:
...
  "/api/chat/{conversation_id}":
    get:
      tags:
      - chat
      summary: Get Conversation
      operationId: get_conversation_api_chat__conversation_id__get
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ChatResponse"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    delete:
      tags:
      - chat
      summary: Delete Conversation
      operationId: delete_conversation_api_chat__conversation_id__delete
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    ChatMessageResponse:
      properties:
        role:
          type: string
          title: Role
        content:
          type: string
          title: Content
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
      type: object
      required:
      - role
      - content
      title: ChatMessageResponse
...
    ChatResponse:
      properties:
        conversation_id:
          type: string
          title: Conversation Id
        message:
          "$ref": "#/components/schemas/ChatMessageResponse"
        history:
          items:
            "$ref": "#/components/schemas/ChatMessageResponse"
          type: array
          title: History
      type: object
      required:
      - conversation_id
      - message
      - history
      title: ChatResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            "$ref": "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError